basic usermode code execution
The main thing here is the introduction of proc_run_code which sets up a
few pages of memory for the supplied code and its stack in a new page
directory and then jumps to this code in usermode. For now, the code
that runs just sets `eax` to 42 and loops.
I introduced preliminary structures to manage processes, mainly
process_t.
I needed to allocate 4 MiB of physical memory aligned on a 4 MiB
boundary, and in doing so I found and fixed a bug in my physical memory
manager.

+++ kernel/include/kernel/gdt.h
@@ -50,7 +50,7 @@ typedef struct {
 void init_gdt();
 void gdt_set_entry(uint32_t num, uint32_t base, uint32_t limit, uint8_t access, uint8_t granularity);
 void gdt_write_tss(uint32_t num, uint32_t ss0, uint32_t esp0);
-void gdt_set_kernel_stack(uint32_t stack);
+void gdt_set_kernel_stack(uintptr_t stack);
 
 extern void gdt_load(uintptr_t gdt_ptr);
